+++ src/common/p25/lc/tsbk/OSP_SNDCP_CH_GNT.cpp
@@ -50,10 +50,10 @@ void OSP_SNDCP_CH_GNT::encode(uint8_t* data, bool rawTSBK, bool noTrellis)
     ulong64_t tsbkValue = 0U;
 
     uint32_t calcSpace = (uint32_t)(m_siteIdenEntry.chSpaceKhz() / 0.125);
-    float calcTxOffset = m_siteIdenEntry.txOffsetMhz() * 1000000;
+    float calcTxOffset = m_siteIdenEntry.txOffsetMhz() * 1000000.0;
 
     uint32_t txFrequency = (uint32_t)((m_siteIdenEntry.baseFrequency() + ((calcSpace * 125) * m_dataChannelNo)));
-    uint32_t rxFrequency = (uint32_t)(txFrequency + calcTxOffset);
+    uint32_t rxFrequency = (uint32_t)(txFrequency + (uint32_t)calcTxOffset);
 
     uint32_t rootFreq = rxFrequency - m_siteIdenEntry.baseFrequency();
     uint32_t rxChNo = rootFreq / (m_siteIdenEntry.chSpaceKhz() * 1000);
